requirements.txt
numpy==1.19.4
Pillow==8.0.1
CairoSVG==2.5.0
